requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0

# NLP & Transformers
transformers>=4.30.0
datasets>=2.14.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
facenet-pytorch>=2.5.0
mtcnn>=0.1.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Explainability
shap>=0.42.0
captum>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
streamlit>=1.25.0

# API & Downloads
kaggle>=1.5.16
huggingface-hub>=0.16.0
requests>=2.31.0

# Video Processing
ffmpeg-python>=0.2.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
